cadastrando usuario e aplicando login, no data base a senha fica criptografada, alem de criar outros arquivos para autenticar as paginas em php

+++ delete.php
@@ -1,31 +1,44 @@
 <?php
 
+session_start();
+if(!isset($_SESSION['user_id'])) {
+    header('Location: login.php');
+    exit();
+}
+
 require_once 'conn.php';
 
 try {
-    if(isset($_GET['id'])){
+    if (isset($_GET['id'])) {
         $id = $_GET['id'];
 
         $sql = "DELETE FROM crud_php WHERE id = ?";
         $stmt = $conn->prepare($sql);
 
-        if($stmt) {
+        if ($stmt) {
             $stmt->bind_param("i", $id);
             if ($stmt->execute()) {
+                session_start();
+                $_SESSION['message'] = "Tarefa deletada com sucesso!";
+                $_SESSION['message_type'] = "success";
                 header("Location: index.php");
                 exit();
             } else {
+                session_start();
+                $_SESSION['message'] = "Erro ao deletar a tarefa.";
+                $_SESSION['message_type'] = "danger";
                 throw new Exception("Erro ao executar a exclusão: " . $stmt->error);
             }
+
             $stmt->close();
         } else {
-            throw new Exception("erro ao preparar a consulta: " . $conn->error);
+            throw new Exception("Erro ao preparar a consulta: " . $conn->error);
         }
     } else {
         throw new Exception("ID da tarefa não fornecido.");
     }
 } catch (Exception $e) {
     echo "Erro: " . $e->getMessage();
-} finally{
+} finally {
     $conn->close();
 }
